refer #2683 
- First step to introduce checksum metadata leveraging the checksum calculation to S3.
- `aws-smithy-http`: Needed because of `SdkBody` see https://github.com/awslabs/smithy-rs/issues/1759
- `aws-types`: Needed because of `SharedCredentialsProvider`, the recommended way from aws is something like `aws_config::from_env().region("us-east-1").load().await` but that is problematic because of:

    - `sync -> async ` in the creation of S3Client and i don't want to change the signature of any method in this class.
    - We do not need the four default steps in https://github.com/awslabs/aws-sdk-rust/blob/main/sdk/aws-config/src/default_provider/credentials.rs#L235

- `Hyper`: Similar to what's currently doing Rusoto in https://github.com/rusoto/rusoto/blob/master/rusoto/signature/src/signature.rs#L59 to stream the body, see also https://github.com/awslabs/aws-sdk-rust/discussions/361 

Need to confirm all tests work in the pipeline and check compilation times outside my dev environment.