@SomeoneToIgnore 2022-11-14
Had tested it on a VM and local mac, and saw no difference in compilation times:
```
time env RUSTC_WRAPPER="" cargo build --all --timings [--release]
-------------

Mac

old: 
    Finished dev [optimized + debuginfo] target(s) in 2m 50s
env RUSTC_WRAPPER="" cargo build --all --timings  1438.32s user 58.24s system 876% cpu 2:50.81 total

old, release:
    Finished release [optimized + debuginfo] target(s) in 2m 04s
env RUSTC_WRAPPER="" cargo build --all --release --timings  1021.31s user 42.76s system 856% cpu 2:04.28 total

new: 
    Finished dev [optimized + debuginfo] target(s) in 2m 48s
env RUSTC_WRAPPER="" cargo build --all --timings  1462.21s user 59.64s system 902% cpu 2:48.57 total

new, release:
    Finished release [optimized + debuginfo] target(s) in 2m 16s
env RUSTC_WRAPPER="" cargo build --all --release --timings  1180.12s user 47.42s system 897% cpu 2:16.79 total

-------------

Linux

old:
    Finished dev [optimized + debuginfo] target(s) in 6m 39s
env RUSTC_WRAPPER="" cargo build --all --timings  2978.42s user 111.40s system 773% cpu 6:39.22 total

old, release:
    Finished release [optimized + debuginfo] target(s) in 4m 40s
env RUSTC_WRAPPER="" cargo build --all --release --timings  2085.85s user 79.98s system 771% cpu 4:40.78 total

new: 
    Finished dev [optimized + debuginfo] target(s) in 6m 41s
env RUSTC_WRAPPER="" cargo build --all --timings  3012.22s user 112.60s system 777% cpu 6:41.83 total

new, release:
    Finished release [optimized + debuginfo] target(s) in 4m 41s
env RUSTC_WRAPPER="" cargo build --all --release --timings  2100.62s user 80.31s system 775% cpu 4:41.35 total
```