@SomeoneToIgnore 2022-11-14
I've attached the timing profiles of the release builds:

Interesting, that despite the build time not changing, new S3-related crate takes longer to compile generally.
<img width="1728" alt="image" src="https://user-images.githubusercontent.com/2690773/201713567-6c5664f7-e702-45af-a2d7-96a27d30d16f.png">

<img width="1728" alt="image" src="https://user-images.githubusercontent.com/2690773/201713619-17e71b9e-46f0-4231-b62c-e4c70c4fee9f.png">


[cargo_build_timings.zip](https://github.com/neondatabase/neon/files/10004985/cargo_build_timings.zip)