@SomeoneToIgnore 2022-11-14
I see an increased failure rate in the tests, check out the Allure reports, there are logs (pytest, pageserver, safekeeper, pg, etc.) in the attachments per test.

<img width="1728" alt="image" src="https://user-images.githubusercontent.com/2690773/201725496-80b3c234-d816-449c-b29a-1ac272f8446b.png">


Debug: https://neon-github-public-dev.s3.amazonaws.com/reports/pr-2812/debug/3463396532/index.html#categories/8fb3a91ba5aaf9de24cc8a92edc82b5d

Release: https://neon-github-public-dev.s3.amazonaws.com/reports/pr-2812/release/3463396532/index.html#suites/3de20ced56df482626cff043a3f2c817

There's also a docker-compose set-up and test we have in our repo, and while that does not fail explicitly (since PS could work without the remote storage), it shows some interesting errors:

```
docker-compose-pageserver-1        | 2022-11-14T16:50:09.593840Z ERROR storage_sync_loop:storage_sync_loop_step:process_sync_task_batch{sync_id=f75d15f012765a22bdda57cdeea17eda/9d2ae9dcaea4e1aff81d7ed6be860ec8}:upload_timeline_data: Failed to upload a layer for timeline f75d15f012765a22bdda57cdeea17eda/9d2ae9dcaea4e1aff81d7ed6be860ec8: Failed to upload layer file for f75d15f012765a22bdda57cdeea17eda/9d2ae9dcaea4e1aff81d7ed6be860ec8
[685](https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:686)
docker-compose-pageserver-1        | 
[686](https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:687)
docker-compose-pageserver-1        | Caused by:
[687](https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:688)
docker-compose-pageserver-1        |     0: Failed to upload from '/data/.neon/tenants/f75d15f012765a22bdda57cdeea17eda/timelines/9d2ae9dcaea4e1aff81d7ed6be860ec8/000000000000000000000000000000000000-FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF__0000000001696070-00000000016960E9' to storage path '"pageserver/tenants/f75d15f012765a22bdda57cdeea17eda/timelines/9d2ae9dcaea4e1aff81d7ed6be860ec8/000000000000000000000000000000000000-FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF__0000000001696070-00000000016960E9"'
[688](https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:689)
docker-compose-pageserver-1        |     1: failed to construct request: error during request construction: Only request bodies with a known size can be checksum validated.
[689](https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:690)
docker-compose-pageserver-1        |     2: error during request construction: Only request bodies with a known size can be checksum validated.
[690](https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:691)
docker-compose-pageserver-1        |     3: Only request bodies with a known size can be checksum validated.
[691](https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:692)
```

https://github.com/neondatabase/neon/actions/runs/3463396532/jobs/5783718923#step:4:685