@andresrsanchez 2022-11-14
Thanks for the review @SomeoneToIgnore! 
Yep i trimmed the `default-features` from aws-config because initially the results were much worse... 

I will remove all the checksum thing and check the errors, thanks!!!!