@andresrsanchez 2022-11-15
refer: https://github.com/awslabs/aws-sdk-rust/issues/668